{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "fastapi-{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "This is {{ cookiecutter.project_slug }}"
authors = ["ha.hao.minh <ha.hao.minh@sun-asterisk.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.36"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
loguru = "^0.6.0"
alembic = "^1.7.7"
PyYAML = "^6.0"
python-dotenv = "^0.20.0"
python-jose = "^3.3.0"
gunicorn = "^20.1.0"
requests = "^2.27.1"
Faker = "^13.11.1"
coverage = "^6.3.3"
celery = "^5.2.6"
rabbitmq = "^0.2.0"
redis = "^4.3.1"
psycopg2-binary = "^2.9.3"
sentry-sdk = "^1.5.12"
boto3 = "^1.23.6"
python-multipart = "^0.0.5"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
flower = "^1.0.0"
python-logstash-async = "^2.5.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.9"
pytest = "^7.1.2"
pytest-mock = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
